Move DynamoClientModule into the new package
