+++ catalogue_pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/Server.scala
@@ -7,6 +7,7 @@ import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
+import uk.ac.wellcome.storage.dynamo.DynamoClientModule
 import uk.ac.wellcome.platform.recorder.modules.RecorderModule
 
 object ServerMain extends Server
