+++ reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex_worker/Server.scala
@@ -13,6 +13,7 @@ import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.platform.reindex_worker.modules.ReindexerWorkerModule
+import uk.ac.wellcome.storage.dynamo.DynamoClientModule
 
 object ServerMain extends Server
