+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/dynamo/DynamoClientModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.finatra.modules
+package uk.ac.wellcome.storage.dynamo
 
 import javax.inject.Singleton
 
@@ -8,6 +8,7 @@ import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
 import uk.ac.wellcome.models.aws.AWSConfig
+import uk.ac.wellcome.finatra.modules.AWSConfigModule
 import com.amazonaws.services.dynamodbv2._
 
 object DynamoClientModule extends TwitterModule {
